requirements.txt
pandas==2.3.1
numpy==1.26.4
scikit-learn==1.7.1
fastapi==0.117.1
uvicorn==0.40.0
requests==2.32.3
mlflow==3.11.1
dvc==3.66.1
apache-airflow==3.0.3
prometheus-client==0.22.1
joblib==1.5.1
pytest==8.4.1
httpx==0.28.1
python-multipart==0.0.20
jinja2==3.1.6

